Firewall CMR cleanup
