+++ packages/manager/src/features/Firewalls/FirewallDetail/Devices/FirewallDeviceRow.tsx
@@ -1,8 +1,19 @@
 import * as React from 'react';
+import { Link } from 'react-router-dom';
 import { compose } from 'recompose';
-import TableCell from 'src/components/TableCell';
-import TableRow from 'src/components/TableRow';
+import { makeStyles } from 'src/components/core/styles';
+import TableCell from 'src/components/TableCell/TableCell_CMR';
+import TableRow from 'src/components/TableRow/TableRow_CMR';
 import ActionMenu, { Props as ActionProps } from './FirewallDeviceActionMenu';
+
+const useStyles = makeStyles(() => ({
+  actionCell: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'flex-end',
+    padding: 0
+  }
+}));
 
 interface Props {
   entityID: number;
@@ -11,16 +22,20 @@ interface Props {
 export type CombinedProps = Props & ActionProps;
 
 export const FirewallDeviceRow: React.FC<CombinedProps> = props => {
+  const classes = useStyles();
   const { deviceLabel, deviceID, entityID } = props;
 
   return (
     <TableRow
-      rowLink={`/linodes/${entityID}`}
       data-testid={`firewall-device-row-${deviceID}`}
       ariaLabel={`Device ${deviceLabel}`}
     >
-      <TableCell>{deviceLabel}</TableCell>
       <TableCell>
+        <Link to={`/linodes/${entityID}`} tabIndex={0}>
+          {deviceLabel}
+        </Link>
+      </TableCell>
+      <TableCell className={classes.actionCell}>
         <ActionMenu {...props} />
       </TableCell>
     </TableRow>
