+++ packages/manager/src/features/Firewalls/index.tsx
@@ -2,13 +2,10 @@ import * as React from 'react';
 import { Route, RouteComponentProps, Switch } from 'react-router-dom';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import SuspenseLoader from 'src/components/SuspenseLoader';
-import useFlags from 'src/hooks/useFlags';
 import useReduxLoad from 'src/hooks/useReduxLoad';
 
 const FirewallLanding = React.lazy(() => import('./FirewallLanding'));
-const FirewallLanding_CMR = React.lazy(() =>
-  import('./FirewallLanding/FirewallLanding_CMR')
-);
+
 const FirewallDetail = React.lazy(() => import('./FirewallDetail'));
 
 type Props = RouteComponentProps<{}>;
@@ -20,7 +17,6 @@ const Firewall: React.FC<CombinedProps> = props => {
     match: { path }
   } = props;
 
-  const flags = useFlags();
   useReduxLoad(['firewalls']);
 
   return (
@@ -28,11 +24,7 @@ const Firewall: React.FC<CombinedProps> = props => {
       <React.Fragment>
         <DocumentTitleSegment segment="Firewalls" />
         <Switch>
-          <Route
-            exact
-            path={`${path}(/create)?`}
-            component={flags.cmr ? FirewallLanding_CMR : FirewallLanding}
-          />
+          <Route exact path={`${path}(/create)?`} component={FirewallLanding} />
           <Route path={`${path}/:id`} component={FirewallDetail} />
           <Route component={FirewallLanding} />
         </Switch>
